Divided class into packages
