--- src/main/java/ListCommand.java
@@ -1,12 +1,0 @@
-public class ListCommand extends Command {
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        ui.showList();
-    }
-}
