+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,8 +1,17 @@
-public class MarkCommand extends Command {
+package duke.command;
+
+import duke.Message;
+import duke.DukeException;
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class DeleteCommand extends Command {
 
     private final int index;
 
-    public MarkCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
 
@@ -14,12 +23,11 @@ public class MarkCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         try {
-            tasks.markTaskAtPos(this.index);
-            Task currentTask = tasks.getTask(this.index);
+            Task deletedTask = tasks.deleteTaskAtPos(this.index);
             storage.save(tasks);
-            ui.showMarked(currentTask);
+            ui.showDeleted(deletedTask);
         } catch (IndexOutOfBoundsException e) {
-            if (tasks.getCount() == 0) {
+            if (tasks.getCount() == 0){
                 throw new DukeException(Message.INVALID_ACCESS_EMPTY_TASKLIST);
             } else {
                 throw new DukeException(Message.returnTaskNotFound(tasks));
