+++ src/main/java/duke/command/MarkCommand.java
@@ -1,8 +1,17 @@
-public class DeleteCommand extends Command {
+package duke.command;
+
+import duke.Message;
+import duke.DukeException;
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class MarkCommand extends Command {
 
     private final int index;
 
-    public DeleteCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
@@ -14,11 +23,12 @@ public class DeleteCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         try {
-            Task deletedTask = tasks.deleteTaskAtPos(this.index);
+            tasks.markTaskAtPos(this.index);
+            Task currentTask = tasks.getTask(this.index);
             storage.save(tasks);
-            ui.showDeleted(deletedTask);
+            ui.showMarked(currentTask);
         } catch (IndexOutOfBoundsException e) {
-            if (tasks.getCount() == 0){
+            if (tasks.getCount() == 0) {
                 throw new DukeException(Message.INVALID_ACCESS_EMPTY_TASKLIST);
             } else {
                 throw new DukeException(Message.returnTaskNotFound(tasks));
