+++ src/main/java/duke/ui/Ui.java
@@ -1,3 +1,9 @@
+package duke.ui;
+
+import duke.Message;
+import duke.task.Task;
+import duke.task.TaskList;
+
 import java.util.Scanner;
 
 public class Ui {
